practic/requirements.txt
celery==5.4.0
Django==4.2.14
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory-boy==3.3.0
python-decouple==3.8
requests==2.32.3
redis==4.6.0
psycopg2>=2.8